+++ src/main/java/mil/nga/geopackage/features/user/ManualFeatureQuery.java
@@ -3500,7 +3500,7 @@ public class ManualFeatureQuery {
 	/**
 	 * Manually query for rows within the geometry envelope, starting at the
 	 * offset and returning no more than the limit
-	 * 
+	 * <p>
 	 * WARNING: This method must iterate from the 0 offset each time, is
 	 * extremely inefficient, and not recommended for use
 	 * 
@@ -3534,7 +3534,7 @@ public class ManualFeatureQuery {
 	/**
 	 * Manually query for rows within the bounds, starting at the offset and
 	 * returning no more than the limit
-	 * 
+	 * <p>
 	 * WARNING: This method must iterate from the 0 offset each time, is
 	 * extremely inefficient, and not recommended for use
 	 *