+++ CHANGELOG.md
@@ -4,9 +4,16 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 6.2.2 (TBD)
+## 6.3.1 (TBD)
 
 * TBD
+
+## [6.3.0](https://github.com/ngageoint/geopackage-java/releases/tag/6.3.0) (04-18-2022)
+
+* geopackage-core version 6.3.0
+* OGC Related Tables Extension additional mappings methods
+* NGA Feature Style Extension additional style and icon row methods
+* Feature Style expanded pixel bounds
 
 ## [6.2.1](https://github.com/ngageoint/geopackage-java/releases/tag/6.2.1) (03-11-2022)